dolfinx test tidyup
